requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.2
frozenlist==1.3.3
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
openai==0.27.8
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyforest==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
requests==2.31.0
seaborn==0.12.2
six==1.16.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
zipp==3.15.0
